# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Expense Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Expense {
  id: number;
  amount: number;
  description: string;
  categoryId: number;
  categoryName: string;
  /** @nullable */
  cardId?: number | null;
  /** @nullable */
  cardName?: string | null;
  /** @nullable */
  cardColor?: string | null;
  /** @nullable */
  cardLastFour?: string | null;
  /** @nullable */
  notes?: string | null;
  date: string;
  createdAt: string;
}

export interface ExpenseInput {
  /** @minimum 0.01 */
  amount: number;
  /** @minLength 1 */
  description: string;
  categoryId: number;
  cardId?: number;
  notes?: string;
  date: string;
}

export interface ExpenseUpdate {
  /** @minimum 0.01 */
  amount?: number;
  /** @minLength 1 */
  description?: string;
  categoryId?: number;
  /** @nullable */
  cardId?: number | null;
  /** @nullable */
  notes?: string | null;
  date?: string;
}

export interface Category {
  id: number;
  name: string;
  color: string;
  isDefault: boolean;
}

export interface CategoryInput {
  /** @minLength 1 */
  name: string;
  color: string;
}

export interface Card {
  id: number;
  name: string;
  /** @nullable */
  lastFour?: string | null;
  color: string;
  createdAt: string;
}

export interface CardInput {
  /** @minLength 1 */
  name: string;
  lastFour?: string;
  color: string;
}

export interface CategoryStat {
  categoryId: number;
  categoryName: string;
  color: string;
  total: number;
  count: number;
  percentage: number;
}

export interface CardStat {
  /** @nullable */
  cardId: number | null;
  cardName: string;
  color: string;
  /** @nullable */
  lastFour?: string | null;
  total: number;
  count: number;
  percentage: number;
}

export interface MonthlyStat {
  /** Format: YYYY-MM */
  month: string;
  total: number;
  count: number;
}

export interface ExpenseSummary {
  totalAllTime: number;
  totalThisMonth: number;
  totalLastMonth: number;
  expenseCount: number;
  /** @nullable */
  topCategory: string | null;
  averagePerDay: number;
}

export type ListExpensesParams = {
  category?: string;
  /**
   * Format: YYYY-MM
   */
  month?: string;
};

export type GetStatsByCategoryParams = {
  /**
   * Format: YYYY-MM. Omit for all-time.
   */
  month?: string;
};

export type GetStatsByCardParams = {
  /**
   * Format: YYYY-MM. Omit for all-time.
   */
  month?: string;
};