# lib/api-zod/src/generated/types/card.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Expense Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Card {
  id: number;
  name: string;
  /** @nullable */
  lastFour?: string | null;
  color: string;
  createdAt: Date;
}