# lib/api-zod/src/generated/types/cardInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Expense Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CardInput {
  /** @minLength 1 */
  name: string;
  lastFour?: string;
  color: string;
}