# lib/api-zod/src/generated/types/cardStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Expense Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CardStat {
  /** @nullable */
  cardId: number | null;
  cardName: string;
  color: string;
  /** @nullable */
  lastFour?: string | null;
  total: number;
  count: number;
  percentage: number;
}