# lib/api-zod/src/generated/types/expense.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Expense Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Expense {
  id: number;
  amount: number;
  description: string;
  categoryId: number;
  categoryName: string;
  /** @nullable */
  cardId?: number | null;
  /** @nullable */
  cardName?: string | null;
  /** @nullable */
  cardColor?: string | null;
  /** @nullable */
  cardLastFour?: string | null;
  /** @nullable */
  notes?: string | null;
  date: Date;
  createdAt: Date;
}