# lib/api-zod/src/generated/types/expenseInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Expense Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface ExpenseInput {
  /** @minimum 0.01 */
  amount: number;
  /** @minLength 1 */
  description: string;
  categoryId: number;
  cardId?: number;
  notes?: string;
  date: Date;
}